Write User's externalId to BigQuery
This externalId comes when the client wants to use the user features
and provided a unique id when they authenticated. This is a different
use case from the other case when they are seen only as a service
account but they send a Boclips-User-Id header
in every request for analytic purposes.

[#176052329]
Signed-off-by: Marcin Janik <marcin@boclips.com>